+++ changelog.md
@@ -1,5 +1,7 @@
-# cellseg: MultiClass Cell Segmentation
+# cellseg: Quantify and Classify Cells
 
 **version 0.1.0**
+
+* Implement data loaders. 
 
 * Conceptualize project